Consolidate image tagging and reduce line count
