+++ lib/images/tag_manager.go
@@ -1,27 +1,14 @@
 package images
 
 import (
-	"context"
 	"errors"
-	"fmt"
 	"log/slog"
 	"strings"
 	"time"
-
-	"github.com/kernel/hypeman/lib/paths"
 )
 
-type requestedTagKey struct {
-	repository string
-	tag        string
-}
-
-func tagGenerationKey(repository, tag string) string {
+func tagKey(repository, tag string) string {
 	return repository + ":" + tag
-}
-
-func requestedTagKeyFor(repository, tag string) requestedTagKey {
-	return requestedTagKey{repository: repository, tag: tag}
 }
 
 func (m *manager) ensureTagState() {
@@ -29,19 +16,19 @@ func (m *manager) ensureTagState() {
 		m.tagGenerations = make(map[string]uint64)
 	}
 	if m.requestedTags == nil {
-		m.requestedTags = make(map[requestedTagKey]string)
+		m.requestedTags = make(map[string]string)
 	}
 }
 
 func (m *manager) nextTagGeneration(repository, tag string) uint64 {
 	m.ensureTagState()
-	key := tagGenerationKey(repository, tag)
+	key := tagKey(repository, tag)
 	m.tagGenerations[key]++
 	return m.tagGenerations[key]
 }
 
 func (m *manager) revertTagGeneration(repository, tag string) {
-	key := tagGenerationKey(repository, tag)
+	key := tagKey(repository, tag)
 	if m.tagGenerations[key] <= 1 {
 		delete(m.tagGenerations, key)
 		return
@@ -50,13 +37,13 @@ func (m *manager) revertTagGeneration(repository, tag string) {
 }
 
 func (m *manager) releaseTagGeneration(repository, tag string, generation uint64) {
-	if m.tagGenerations[tagGenerationKey(repository, tag)] == generation {
+	if m.tagGenerations[tagKey(repository, tag)] == generation {
 		m.revertTagGeneration(repository, tag)
 	}
 }
 
 func (m *manager) forgetTagState(repository, tag string) {
-	delete(m.tagGenerations, tagGenerationKey(repository, tag))
+	delete(m.tagGenerations, tagKey(repository, tag))
 	m.clearRequestedTag(repository, tag, "")
 }
 
@@ -65,10 +52,7 @@ func (m *manager) restoreTagState(metas []*imageMetadata) {
 	defer m.createMu.Unlock()
 	m.ensureTagState()
 	for _, meta := range metas {
-		if !isPendingImageStatus(meta.Status) {
-			continue
-		}
-		if meta.Digest == "" {
+		if !isPendingImageStatus(meta.Status) || meta.Digest == "" {
 			continue
 		}
 		digestHex := meta.digestHex()
@@ -89,10 +73,10 @@ func (m *manager) restoreRequestedTag(repository, tag, digestHex string, generat
 	if tag == "" {
 		return
 	}
-	key := tagGenerationKey(repository, tag)
+	key := tagKey(repository, tag)
 	if currentGen, ok := m.tagGenerations[key]; !ok || generation >= currentGen {
 		m.tagGenerations[key] = generation
-		m.requestedTags[requestedTagKeyFor(repository, tag)] = digestHex
+		m.requestedTags[key] = digestHex
 	}
 }
 
@@ -103,9 +87,6 @@ func (m *manager) claimTagForStatus(meta *imageMetadata, ref *ResolvedRef) error
 	if meta.Status == StatusReady {
 		return m.claimReadyTag(ref.Repository(), ref.Tag(), ref.DigestHex())
 	}
-	// Keep the existing tag visible until the build succeeds, but remember this
-	// digest as the newest request for the tag. Finalization will claim it along
-	// with every other tag waiting for the same digest.
 	if meta.RequestedTag == ref.Tag() && strings.HasPrefix(meta.Name, ref.Repository()+":") {
 		wasCanceled := meta.TagClaimCanceled
 		meta.TagClaimCanceled = false
@@ -163,11 +144,11 @@ func (m *manager) claimReadyTag(repository, tag, digestHex string) error {
 
 func (m *manager) trackRequestedTag(repository, tag, digestHex string) {
 	m.ensureTagState()
-	m.requestedTags[requestedTagKeyFor(repository, tag)] = digestHex
+	m.requestedTags[tagKey(repository, tag)] = digestHex
 }
 
 func (m *manager) clearRequestedTag(repository, tag, digestHex string) {
-	key := requestedTagKeyFor(repository, tag)
+	key := tagKey(repository, tag)
 	if current, ok := m.requestedTags[key]; ok && (digestHex == "" || current == digestHex) {
 		delete(m.requestedTags, key)
 	}
@@ -177,8 +158,9 @@ func (m *manager) clearRequestedDigest(digestHex string) {
 	for key, current := range m.requestedTags {
 		if current == digestHex {
 			delete(m.requestedTags, key)
-			if nextDigest := m.findPendingDigestForTag(key.repository, key.tag, digestHex); nextDigest != "" {
-				m.requestedTags[key] = nextDigest
+			repo, tag, _ := strings.Cut(key, ":")
+			if next := m.findPendingDigestForTag(repo, tag, digestHex); next != "" {
+				m.requestedTags[key] = next
 			}
 		}
 	}
@@ -202,18 +184,14 @@ func (m *manager) findPendingDigestForTag(repository, tag, excludeDigestHex stri
 			ref, err := ParseNormalizedRef(meta.Name)
 			if err == nil && ref.Repository() == repository {
 				if meta.TagGeneration >= bestGen || (meta.TagGeneration == bestGen && meta.CreatedAt.After(bestTime)) {
-					bestDigest = meta.digestHex()
-					bestTime = meta.CreatedAt
-					bestGen = meta.TagGeneration
+					bestDigest, bestTime, bestGen = meta.digestHex(), meta.CreatedAt, meta.TagGeneration
 				}
 			}
 		}
 		for _, claim := range meta.TagClaims {
 			if claim.Repository == repository && claim.Tag == tag {
 				if claim.TagGeneration >= bestGen || (claim.TagGeneration == bestGen && meta.CreatedAt.After(bestTime)) {
-					bestDigest = meta.digestHex()
-					bestTime = meta.CreatedAt
-					bestGen = claim.TagGeneration
+					bestDigest, bestTime, bestGen = meta.digestHex(), meta.CreatedAt, claim.TagGeneration
 				}
 			}
 		}
@@ -305,7 +283,7 @@ func cancelTagClaim(meta *imageMetadata, ref *NormalizedRef, repository, tag str
 
 func (m *manager) requestedTagImage(ref *NormalizedRef) *Image {
 	m.createMu.Lock()
-	digestHex, ok := m.requestedTags[requestedTagKeyFor(ref.Repository(), ref.Tag())]
+	digestHex, ok := m.requestedTags[tagKey(ref.Repository(), ref.Tag())]
 	m.createMu.Unlock()
 	if !ok {
 		return nil
@@ -323,17 +301,16 @@ func (m *manager) claimRequestedTags(ref *ResolvedRef, meta *imageMetadata) bool
 	if primaryTag == "" {
 		primaryTag = ref.Tag()
 	}
-	primaryKey := requestedTagKeyFor(ref.Repository(), primaryTag)
+	primaryKey := tagKey(ref.Repository(), primaryTag)
 
-	claimed := false
-	primaryClaimed := false
+	claimed, primaryClaimed := false, false
 	for key, requestedDigest := range m.requestedTags {
 		if requestedDigest != digestHex {
 			continue
 		}
-		repository, tag := key.repository, key.tag
-		if err := createTagSymlink(m.paths, repository, tag, digestHex); err != nil {
-			slog.Warn("failed to claim image tag", "repository", repository, "tag", tag, "digest", digestHex, "error", err)
+		repo, tag, _ := strings.Cut(key, ":")
+		if err := createTagSymlink(m.paths, repo, tag, digestHex); err != nil {
+			slog.Warn("failed to claim image tag", "repository", repo, "tag", tag, "digest", digestHex, "error", err)
 			continue
 		}
 		delete(m.requestedTags, key)
@@ -346,7 +323,7 @@ func (m *manager) claimRequestedTags(ref *ResolvedRef, meta *imageMetadata) bool
 		claimed = true
 	}
 	for _, claim := range meta.TagClaims {
-		claimKey := requestedTagKeyFor(claim.Repository, claim.Tag)
+		claimKey := tagKey(claim.Repository, claim.Tag)
 		if m.tagClaimIsCurrent(claim.Repository, claim.Tag, digestHex, meta, false) {
 			if _, hasReq := m.requestedTags[claimKey]; !hasReq {
 				if err := createTagSymlink(m.paths, claim.Repository, claim.Tag, digestHex); err == nil {
@@ -355,29 +332,7 @@ func (m *manager) claimRequestedTags(ref *ResolvedRef, meta *imageMetadata) bool
 			}
 		}
 	}
-	if claimed {
-		return true
-	}
-	// Digest-only images have no tag claim but must remain addressable by their
-	// digest after finalization.
-	return meta.RequestedTag == "" && ref.Tag() == ""
-}
-
-func (m *manager) claimMatchingTags(digestHex string) bool {
-	claimed := false
-	for key, requestedDigest := range m.requestedTags {
-		if requestedDigest != digestHex {
-			continue
-		}
-		repository, tag := key.repository, key.tag
-		if err := createTagSymlink(m.paths, repository, tag, digestHex); err != nil {
-			slog.Warn("failed to claim image tag", "repository", repository, "tag", tag, "digest", digestHex, "error", err)
-			continue
-		}
-		delete(m.requestedTags, key)
-		claimed = true
-	}
-	return claimed
+	return claimed || (meta.RequestedTag == "" && ref.Tag() == "")
 }
 
 func (m *manager) claimPrimaryTag(ref *ResolvedRef, meta *imageMetadata, digestHex string) bool {
@@ -391,7 +346,7 @@ func (m *manager) claimPrimaryTag(ref *ResolvedRef, meta *imageMetadata, digestH
 	if primaryTag == "" {
 		return false
 	}
-	primaryKey := requestedTagKeyFor(ref.Repository(), primaryTag)
+	primaryKey := tagKey(ref.Repository(), primaryTag)
 	if _, hasRequest := m.requestedTags[primaryKey]; hasRequest || !m.tagClaimIsCurrent(ref.Repository(), primaryTag, digestHex, meta, meta.RequestedTag == "") {
 		return false
 	}
@@ -403,7 +358,7 @@ func (m *manager) claimPrimaryTag(ref *ResolvedRef, meta *imageMetadata, digestH
 }
 
 func (m *manager) tagClaimIsCurrent(repository, tag, digest string, meta *imageMetadata, allowMissing bool) bool {
-	if tag == "" || m.tagGenerations[tagGenerationKey(repository, tag)] != meta.TagGeneration {
+	if tag == "" || m.tagGenerations[tagKey(repository, tag)] != meta.TagGeneration {
 		return false
 	}
 	current, err := resolveTag(m.paths, repository, tag)
@@ -411,116 +366,4 @@ func (m *manager) tagClaimIsCurrent(repository, tag, digest string, meta *imageM
 		return current == digest || current == meta.PreviousTagDigest
 	}
 	return allowMissing && errors.Is(err, ErrNotFound)
-}
-
-// TagImage creates a ready-image tag without pulling or converting content.
-// Cross-repository tags promote content into the shared layout.
-func (m *manager) TagImage(ctx context.Context, source, target string) (*Image, error) {
-	sourceRef, targetRef, err := parseTagReferences(source, target)
-	if err != nil {
-		return nil, err
-	}
-
-	m.createMu.Lock()
-	defer m.createMu.Unlock()
-
-	previousDigest, err := existingTagDigest(m.paths, targetRef)
-	if err != nil {
-		return nil, err
-	}
-	digestHex, meta, err := m.readyTagImage(sourceRef)
-	if err != nil {
-		return nil, err
-	}
-	if err := m.cancelPendingTag(targetRef.Repository(), targetRef.Tag()); err != nil {
-		return nil, fmt.Errorf("cancel pending image tag: %w", err)
-	}
-	if err := m.installTag(sourceRef, targetRef, digestHex, meta); err != nil {
-		return nil, err
-	}
-	m.nextTagGeneration(targetRef.Repository(), targetRef.Tag())
-	m.clearRequestedTag(targetRef.Repository(), targetRef.Tag(), "")
-	m.cleanupReplacedTag(targetRef, previousDigest, digestHex)
-
-	return meta.toImageFor(targetRef.String()), nil
-}
-
-func existingTagDigest(p *paths.Paths, ref *NormalizedRef) (string, error) {
-	digest, err := resolveTag(p, ref.Repository(), ref.Tag())
-	if err == nil || errors.Is(err, ErrNotFound) || errors.Is(err, errInvalidSymlinkTarget) {
-		return digest, nil
-	}
-	return "", fmt.Errorf("resolve existing target tag: %w", err)
-}
-
-func (m *manager) installTag(source, target *NormalizedRef, digest string, meta *imageMetadata) error {
-	if source.Repository() != target.Repository() {
-		if err := promoteImageToContent(m.paths, source.Repository(), digest, meta); err != nil {
-			return fmt.Errorf("promote image to content: %w", err)
-		}
-	}
-	installed, err := installTagSymlink(m.paths, target.Repository(), target.Tag(), digest)
-	if err != nil {
-		return fmt.Errorf("create image tag: %w", err)
-	}
-	if err := writeMetadata(m.paths, target.Repository(), digest, meta); err != nil {
-		if restoreErr := restoreTagSymlink(&installed); restoreErr != nil {
-			return errors.Join(
-				fmt.Errorf("write tagged image metadata: %w", err),
-				fmt.Errorf("restore image tag: %w", restoreErr),
-			)
-		}
-		return fmt.Errorf("write tagged image metadata: %w", err)
-	}
-	return nil
-}
-
-func (m *manager) cleanupUnclaimedImage(ref *ResolvedRef) {
-	if err := removeDigestIfUnreferenced(m.paths, ref.Repository(), ref.DigestHex(), true); err != nil {
-		slog.Warn("failed to collect stale image", "repository", ref.Repository(), "digest", ref.DigestHex(), "error", err)
-	}
-}
-
-func (m *manager) cleanupReplacedTag(ref *NormalizedRef, previousDigest, digestHex string) {
-	if previousDigest == "" || previousDigest == digestHex {
-		return
-	}
-	count, err := countTagsForDigest(m.paths, ref.Repository(), previousDigest)
-	if err != nil {
-		slog.Warn("failed to count tags for replaced image", "repository", ref.Repository(), "digest", previousDigest, "error", err)
-		return
-	}
-	if count > 0 {
-		return
-	}
-	if err := removeDigestIfUnreferenced(m.paths, ref.Repository(), previousDigest, true); err != nil {
-		slog.Warn("failed to collect replaced image content", "repository", ref.Repository(), "digest", previousDigest, "error", err)
-	}
-	m.refreshDiskUsageTotals()
-}
-
-func parseTagReferences(source, target string) (*NormalizedRef, *NormalizedRef, error) {
-	sourceRef, err := ParseNormalizedRef(source)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: invalid source reference: %s", ErrInvalidName, err)
-	}
-	targetRef, err := ParseNormalizedRef(target)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: invalid target reference: %s", ErrInvalidName, err)
-	}
-	if targetRef.IsDigest() {
-		return nil, nil, fmt.Errorf("%w: target must be a tag reference, not a digest", ErrInvalidName)
-	}
-	return sourceRef, targetRef, nil
-}
-
-func (m *manager) readyTagImage(ref *NormalizedRef) (string, *imageMetadata, error) {
-	digestHex, meta, err := resolveRefMetadata(m.paths, ref)
-	if err != nil {
-		return "", nil, err
-	}
-	if meta.Status != StatusReady {
-		return "", nil, fmt.Errorf("%w: %s", ErrImageNotReady, meta.Status)
-	}
-	return digestHex, meta, nil
 }
